requirements.txt
##controller requirements
gunicorn==20.1.0
Jinja2==3.0.1
PyYAML==5.4.1
flask==2.0.1
Flask-WTF==0.15.1
Flask-Bootstrap==3.3.7.1
flask-nav==0.6
celery==5.1.2
redis==3.5.3
snakemake==6.8.0
pytest==6.2.5
omegaconf==2.1.1
graphviz==0.16